backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# LangChain & LangGraph
langchain==0.1.9
langchain-google-genai==0.0.11
langgraph==0.0.26
langchain-core==0.1.27

# Google AI
google-generativeai==0.4.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
pillow==10.2.0
aiofiles==23.2.1

# Logging
structlog==24.1.0
